Fix 5 issues flagged by Copilot PR review
- query_id_is_reverse: use ptrdiff_t loop index to prevent size_t
  underflow past 0 (UB / infinite loop on short strings)
- mmap_array.hpp: add missing <type_traits> for is_trivially_copyable_v
- MmapArray::grow: throw logic_error instead of calling alloc_file_backed("")
  with empty path when data_ is null and file_backed_ is true
- cmd_hits2emi.cpp: add explicit <cmath>, <cstdlib>, <stdexcept> includes
- em_index.cpp: add missing <limits> for std::numeric_limits

+++ include/agp/damage_alignment_evidence.hpp
@@ -76,7 +76,8 @@ inline const DamageEvidenceSub* find_damage_evidence_sub(char target_aa, char qu
 inline bool query_id_is_reverse(std::string_view query_id) {
     // Expected AGP suffix: ..._<+|->_<0..9>
     if (query_id.size() < 4) return false;
-    for (size_t i = query_id.size() - 1; i >= 2; --i) {
+    // Use signed index to avoid size_t underflow past 0
+    for (ptrdiff_t i = static_cast<ptrdiff_t>(query_id.size()) - 1; i >= 2; --i) {
         if (query_id[i] == '_' &&
             query_id[i - 2] == '_' &&
             (query_id[i - 1] == '+' || query_id[i - 1] == '-')) {
